backend/shared/requirements.txt
# Shared dependencies for all microservices
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
redis==5.2.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0,<5.0.0
python-multipart==0.0.20

